Starting to add support for more list markers
Added two variables that contain lists of markers for numbered and unordered
lists. For the numbered lists the current syntax is `#` representing the list
item number, and all other characters are literal. For unordered lists there is
no special syntax, all characters are literal.

The current method of list detection will need to be changed in order to support
this update.