Made sure markers are matched without magic
This ensures that unless specifically escaped all characters in the markers will
be matched literally.

Note: when substituting the regex in for numbered markers two pairs of
backslashes were necessary (`\\\\d\\\\+`) to get the desired result of `\d\+`
instead of the expected one (`\\d\\+"). I believe this is because the string was
parsed twice by substitute but I am not sure.