Fix issue with AutolistReturn
AutolistReturn had been starting insert mode at the wrong position in
the line. This adds the motion to jump to the beginning of the paste in
order to leave the cursor at the correct position.

This also removes the useless 'a' from normal mode commands in each of
these cases. Returning to insert mode is handled by `startinsert` not
the 'a'.