Fix numbered markers not working after last update
When the plugin was linted and double quoted strings were changed into
single quoted strings the backslashes that are necessary to make the
numbered markers work were removed. That is, a backslash must be put
into the variable so that it can be used again as a pattern later, so
'\\' needs to be written in the string to insert a literal backslash
after the `substitute()` call.