Combine ContinueUp and ContinueDown functions
These functions were essentially the same minus two lines, both
dependent on the continuation direction. This adds `ContinueList(dir)`
and updates the values of 's:dir_up' and 's:dir_down' so that they can
be added to the current line number to get the 'previous' line to check.

The new `ContinueList(dir)` functions the same as the old `ContinueUp()`
and `ContinueDown()` did, nothing the user interacts with should have
changed.