Partial Revert of '95c7f042d8e99d8831c39468e523720d235569cd'
Finding the cursor position in insert mode is just too unreliable. This
goes back to the normal mode mapping for `<Plug>AutolistReturn`. This is
a breaking change for anyone that moved over to the insert mode mapping
(probably just me, but still).

The reason it's a partial revert is to keep the updates to the help
page.