Add buffer-local list markers
This adds the possibility for buffer-local list markers by adding
'b:autolist_numbered_markers' and 'b:autolist_unordered_markers'. These
can be checked in addition to global markers or can be set to override
the global markers with the newly added
'g:autolist_override_global_markers' and
'b:autolist_override_global_markers'.

This also includes a fix for any markers that may contain backslashes
('\'). When defining the marker to be checked two backslashes are
necessary so that the marker itself can be used as a regular expression
in order to detect the list, but only one backslash is intended for
inserting. This is handled by replacing pairs of backslashes before the
list marker is returned for inserting.