Move functions to autoload directory
This improves Vim startup time with this plugin installed, it doesn't
need to source all the functions upfront, especially when they may not
even be used in the session.

Now all functions are in autoload, the only downside is this removes
marker validation. However, after the addition of buffer-local markers
the validation needed updating anyway, since it could only check global
markers when being sourced.