Add simple working renumber feature
Related to #4

This will run an increment on the rest of the current paragraph when
inserting a line above. This works when the first number on each line is
the list item number.

This is broken for unordered lists that have numbers anywhere in the
list. Actual numbered list detection is needed to fix this case.

This also clobbers the t and b marks, it should save any previous values
and reset them after it is finished.